Stop using success count to determine if ES is down
Success count was being used to determine if the server was down.
However, I think we cannot infer anything from that value, given that we
don't know the original number of requests.  So, I think it is best to
only rely on the Failure or Success of the bulk insert Future.